Changed directory structure, refactored type extraction functions
